requirements.txt
# ── Kafka ─────────────────────────────────────────────
kafka-python==2.2.2

# ── 데이터 생성 ────────────────────────────────────────
faker==24.0.0

# ── AWS ───────────────────────────────────────────────
boto3==1.34.0

# ── 공통 ──────────────────────────────────────────────
python-dotenv==1.0.0
pandas==2.0.3


# 전체 데이터 다운로드 시 추가
# pip install kaggle